Refactor Headline and Info to smaller components
